modifié :         The Longest Journey/README.md 	modifié :         The Longest Journey/The Longest Journey - GOG - ResidualVM win32.txt 	modifié :         The Longest Journey/The Longest Journey - GOG - ResidualVM.txt 	modifié :         The Longest Journey/clean.sh

+++ The Longest Journey/clean.sh
@@ -1,9 +1,33 @@
 #!/bin/bash
+
+lang=$1
+echo $lang
+
+if [ -d "drive_c" ]; then 
 cd drive_c
+fi
+
 mkdir "GOG Games"
 mv "The Longest Journey" "GOG Games"
 cd "GOG Games/The Longest Journey"
 rm -f -r app commonappdata __redist station_fix tmp gog* *.dll TLJRegFix.* dxsetup.ini
-cd ../../ResidualVM/
+cd ../..
+
+if [ -d "ResidualVM" ]; then 
+cd ResidualVM
 mv app/* .
 rm -f -r app userappdata
+cd ..
+fi
+
+
+if [ "$lang" == "en" ]; then
+echo "english version is already installed (default)"
+exit
+else
+cd tmp/$lang
+echo installing $lang files...
+cp -R * "../../GOG Games/The Longest Journey"
+fi
+cd ../..
+rm -f -r tmp
